further cleanup of PMTilesElevationProvider and less public
