+++ core/src/main/java/com/graphhopper/GraphHopper.java
@@ -962,7 +962,7 @@ public class GraphHopper {
             maxSpeedCalculator.createDataAccessForParser(baseGraph.getDirectory());
     }
 
-    public static void sortGraphAlongHilbertCurve(BaseGraph graph) {
+    private static void sortGraphAlongHilbertCurve(BaseGraph graph) {
         logger.info("sorting graph along Hilbert curve...");
         StopWatch sw = StopWatch.started();
         NodeAccess na = graph.getNodeAccess();
@@ -990,7 +990,7 @@ public class GraphHopper {
         sortGraphForGivenOrdering(graph, newNodesByOldNodes, newEdgesByOldEdges);
     }
 
-    public static void sortGraphForGivenOrdering(BaseGraph baseGraph, IntArrayList newNodesByOldNodes, IntArrayList newEdgesByOldEdges) {
+    private static void sortGraphForGivenOrdering(BaseGraph baseGraph, IntArrayList newNodesByOldNodes, IntArrayList newEdgesByOldEdges) {
         if (!ArrayUtil.isPermutation(newEdgesByOldEdges))
             throw new IllegalStateException("New edges: not a permutation");
         if (!ArrayUtil.isPermutation(newNodesByOldNodes))
@@ -1004,7 +1004,7 @@ public class GraphHopper {
         logger.info("sorting {} nodes took: {}", Helper.nf(newNodesByOldNodes.size()), sw.stop().getTimeString());
     }
 
-    public static long latLonToHilbertIndex(double lat, double lon, int order) {
+    private static long latLonToHilbertIndex(double lat, double lon, int order) {
         double nx = (lon + 180) / 360;
         double ny = (90 - lat) / 180;
         long size = 1L << order;
@@ -1015,7 +1015,7 @@ public class GraphHopper {
         return xy2d(order, x, y);
     }
 
-    public static long xy2d(int n, long x, long y) {
+    private static long xy2d(int n, long x, long y) {
         long d = 0;
         for (long s = 1L << (n - 1); s > 0; s >>= 1) {
             int rx = (x & s) > 0 ? 1 : 0;
